[refactor] Problem solver directories names
