+++ problem-solver/cxx/example-module/agents/IsomorphicSearchAgent.cpp
