+++ problem-solver/cxx/example-module/agents/IsomorphicSearchAgent.hpp
