+++ problem-solver/cxx/example-module/agents/PathFindingAgent.cpp
