+++ problem-solver/cxx/example-module/agents/SubdividingSearchAgent.hpp
