+++ problem-solver/cxx/example-module/data-structures/ConnectorInfo.hpp
