+++ problem-solver/cxx/example-module/data-structures/ConnectorTemplateInfo.hpp
