+++ problem-solver/cxx/example-module/data-structures/PathInfo.hpp
