+++ problem-solver/cxx/example-module/data-structures/WeightTemplateInfo.hpp
