+++ problem-solver/cxx/example-module/exampleModule.hpp
