+++ problem-solver/cxx/example-module/searcher/PathSearcher.cpp
