+++ problem-solver/cxx/example-module/specifications/agent_of_isomorphic_search/lib_component_ui_menu_isomorphic_search.scs
