+++ problem-solver/cxx/example-module/specifications/agent_of_isomorphic_search/ui_menu_isomorphic_search.scs
