+++ problem-solver/cxx/example-module/specifications/agent_of_subdividing_search/content/response.png
