+++ problem-solver/cxx/example-module/specifications/agent_of_subdividing_search/desc_agent_of_subdividing_search.md
