+++ problem-solver/cxx/example-module/specifications/agent_of_subdividing_search/lib_component_ui_menu_subdividing_search.scs
