+++ problem-solver/cxx/example-module/specifications/agent_of_subdividing_search/ui_menu_subdividing_search.scs
