+++ problem-solver/cxx/example-module/test/test-structures/checkTemplates.scs
