+++ problem-solver/cxx/example-module/test/test-structures/graphWithThreeStepsShortestPath.scs
