+++ problem-solver/cxx/example-module/test/test-structures/graphWithTwoStepsShortestPath.scs
