+++ problem-solver/cxx/example-module/test/test-structures/templates.scs
